requirements.txt
python-dotenv==1.0.0
pydantic==2.5.2
google-generativeai==0.3.1
asyncio==3.4.3 